feature/submission v3

+++ kelas-intermediate/Submission/openmusic/migrations/1663052808601_create-table-albums.js
@@ -9,7 +9,7 @@ exports.up = (pgm) => {
       notNull: true,
     },
     year: {
-      type: 'INTEGER',
+      type: 'SMALLINT',
       notNull: true,
     },
   });
